+++ RELEASE_NOTES.md
@@ -8,6 +8,8 @@
 
 - The SDK is now using the microgrid API client from [`frequenz-client-microgrid`](https://github.com/frequenz-floss/frequenz-client-microgrid-python/). You should update your code if you are using the microgrid API client directly.
 
+- The minimum required `frequenz-channels` version is now [`v1.0.0-rc1`](https://github.com/frequenz-floss/frequenz-channels-python/releases/tag/v1.0.0-rc.1).
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->